refactor(results): shared top + swipeable lower with a rolling word reel
Reworks the results screen so the hero, board, and standings are a single fixed
top shared by both views, and only the lower half swipes — the board no longer
duplicates across panels, the swipe feels anchored, and it sets up standings /
compare-timeline sharing.

- Lifts the replay controller into a useTimelineReplay hook that owns the
  timeline model + playback and derives the shared board highlight and the
  climbing hero. The active lower view decides who drives the shared top: the
  playhead on Timeline, taps/compare on Results.
- Replays now default to the end state, so entering the Timeline view already
  matches the Results view (final board + score, no jump); pressing play
  restarts from 0.
- Replaces the timeline word list with a WordReel: the current find sits large
  and centred, the previous rides smaller above and the next peeks faintly
  below, rolling up one row as the playhead advances (the scrubber keeps the
  overview + seek). TimelineLower packages the reel + scrubber + transport.
- ResultsCarousel now wraps only the lower panels and reports the active index
  up. Removes the old Timeline / ReplayWordList components.

Verified in-browser: Results view unchanged; Timeline shares the top and, on
realtime playback, the hero climbs while the shared board replays the path and
the reel rolls through prev/current/next.
